// src/app/models/doctor.model.ts
export class Doctor{
    name: string;
    dob: any = '';
    phone: any = '';
    online_fees: any = '';
    offline_fees: any = '';
    gender: string = '';
    experience: any = '';
    designation: any = '';
    whatsapp_number: any = '';
    qualifications: string[];
    specializations: string[];
    departments:string = '';

    isAmountSame: boolean;
    isWhatsAppNoSameAsPhone: boolean;
}